+++ dhctl/pkg/config/deckhouse_config_test.go
@@ -22,6 +22,50 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
+const configOverridesTemplate = `
+apiVersion: deckhouse.io/v1
+kind: ClusterConfiguration
+clusterType: Static
+podSubnetCIDR: 10.111.0.0/16
+serviceSubnetCIDR: 10.222.0.0/16
+kubernetesVersion: "1.28"
+clusterDomain: "cluster.local"
+---
+apiVersion: deckhouse.io/v1
+kind: InitConfiguration
+deckhouse:
+  devBranch: aaaa
+{{- if .bundle }}
+  bundle: {{ .bundle }}
+{{- end }}
+
+{{- if .logLevel }}
+  logLevel: {{ .logLevel }}
+{{- end }}
+
+{{- if .configOverrides}}
+{{- .configOverrides | nindent 2 }}
+{{- end }}
+---
+apiVersion: deckhouse.io/v1alpha1
+# type of the configuration section
+kind: StaticClusterConfiguration
+# address space for the cluster's internal network
+internalNetworkCIDRs:
+- 192.168.199.0/24
+---
+{{- if .moduleConfigs}}
+{{- .moduleConfigs }}
+{{- end }}
+`
+
+func assertModuleConfig(t *testing.T, mc *ModuleConfig, enabled bool, version int, setting map[string]interface{}) {
+	require.NotNil(t, mc.Spec.Enabled)
+	require.Equal(t, *mc.Spec.Enabled, enabled)
+	require.Equal(t, mc.Spec.Version, version)
+	require.Equal(t, mc.Spec.Settings, SettingsValues(setting))
+}
+
 func generateMetaConfigForDeckhouseConfigTest(t *testing.T, data map[string]interface{}) *MetaConfig {
 	return generateMetaConfig(t, configOverridesTemplate, data, false)
 }
